1st round of cleanup
